move proto_timestamp to dedicated mod
